requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
minio
pillow
ultralytics  # YOLOv8 (includes torch)
easyocr  # Optimized OCR model
python-multipart
openai  # Used for Groq API (OpenAI-compatible SDK)

# Streaming
sse-starlette  # Server-Sent Events for FastAPI

# HTTP requests for TTS
requests

# Testing
pytest
pytest-asyncio
httpx  # For FastAPI TestClient
python-dotenv
